update template content
