+++ html/options.html
@@ -19,5 +19,5 @@
       </section>
     </main>
   </body>
-  <script src="../src/options.js"></script>
+  <script src="../options.js"></script>
 </html>
